+++ CHANGES.md
@@ -1,3 +1,4 @@
+* The UMI deduplication/correction specified in --soloUMIdedup is used for statistics output, filtering and UB tag in BAM output.
 * Implemented input from SAM/BAM for STARsolo, with options --soloInputSAMattrBarcodeSeq --soloInputSAMattrBarcodeQual to specify SAM tags for barcode read sequence and qualities.
 * Fixed an issue that was causing slighlty incorrect values of Implemented input from SAM/BAM for STARsolo. Fixed an issue that was causing slighlty understimated value of Q30 'Bases in RNA read' in Solo.out/Gene/Summary.csv
 * Implemented --soloCellFilter EmptyDrops_CR option for cell filtering (calling) nearly identical to that of CellRanger 3 and 4.